+++ src/shared/meson.build
@@ -132,6 +132,8 @@ shared_sources = files('''
         nscd-flush.h
         nsflags.c
         nsflags.h
+        numa-util.c
+        numa-util.h
         os-util.c
         os-util.h
         output-mode.c
